--- .github/workflows/test.yml
@@ -1,19 +1,0 @@
-name: Node CI
-
-on: [pull_request, push]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - name: Use Node.js 12.x
-      uses: actions/setup-node@v1
-      with:
-        node-version: 12.x
-    - name: npm install, build, and test
-      run: |
-        npm ci
-        npm test
-      env:
-        CI: true
