+++ tsconfig.json
@@ -6,9 +6,8 @@
     "moduleResolution": "node",
     "declaration": true,
     "outDir": "dist",
+    "lib": ["dom", "dom.iterable", "es2020"],
     "removeComments": true
   },
-  "files": [
-    "src/index.ts"
-  ]
+  "files": ["src/index.ts"]
 }
